requirements.txt
torch~=2.0.1
torchvision~=0.15.1
numpy~=1.24.3
matplotlib~=3.7.1
scipy~=1.10.1
Pillow~=9.5.0
pandas~=2.0.1
opencv-python~=4.7.0.72
PyYAML~=6.0
pamly~=0.1.8
tqdm~=4.66.4
timm~=1.0.3
seaborn~=0.13.2
scikit-learn~=1.5.2
umap-learn~=0.5.7
